+++ src/lib/ast/math/expression.ts
@@ -295,6 +295,11 @@ function doEvaluate(
         }
     }
 
+    if (op == EnumToken.Div && v2 == 0) {
+        
+        return defaultReturn;
+    }
+
     // @ts-ignore
     const val: number | FractionToken = compute(v1, v2, op);