+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# v1.6.2
+
+- [x] fix base64 encoding error when inline sourcemap is enabled
+- [ ] support color interpolation in gradient functions 
+- [x] handle dividing by zero in calc()
+
 # v1.6.1
 
 - [x] fix bugs when computing math functions.